+++ test_deproxy.py
@@ -28,8 +28,7 @@ class TestDefaultHandler(unittest.TestCase):
     def setUp(self):
         self.deproxy_port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.end_point = self.deproxy.add_endpoint(('localhost',
-                                                    self.deproxy_port))
+        self.end_point = self.deproxy.add_endpoint(self.deproxy_port)
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()
@@ -44,8 +43,7 @@ class TestEchoHandler(unittest.TestCase):
     def setUp(self):
         self.deproxy_port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.end_point = self.deproxy.add_endpoint(('localhost',
-                                                    self.deproxy_port))
+        self.end_point = self.deproxy.add_endpoint(self.deproxy_port)
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()
@@ -55,7 +53,7 @@ class TestEchoHandler(unittest.TestCase):
         mc = self.deproxy.make_request('http://localhost:%i/' %
                                        self.deproxy_port, headers=headers,
                                        request_body='this is the body',
-                                       handler_function=deproxy.echo_handler)
+                                       default_handler=deproxy.echo_handler)
         self.assertEquals(int(mc.received_response.code), 200)
         self.assertIn('x-header', mc.received_response.headers)
         self.assertEquals(mc.received_response.headers['x-header'], '12345')
@@ -66,18 +64,17 @@ class TestDelayHandler(unittest.TestCase):
     def setUp(self):
         self.deproxy_port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.end_point = self.deproxy.add_endpoint(('localhost',
-                                                    self.deproxy_port))
+        self.end_point = self.deproxy.add_endpoint(self.deproxy_port)
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()
 
     def test_delay_handler(self):
-        handler = deproxy.delay(3, deproxy.default_handler)
+        handler = deproxy.delay(3, deproxy.simple_handler)
         t1 = time.time()
         mc = self.deproxy.make_request('http://localhost:%i/' %
                                        self.deproxy_port,
-                                       handler_function=handler)
+                                       default_handler=handler)
         t2 = time.time()
         self.assertEquals(int(mc.received_response.code), 200)
         self.assertGreaterEqual(t2 - t1, 3)
@@ -88,16 +85,16 @@ class TestRoute(unittest.TestCase):
     def setUp(self):
         self.deproxy_port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.end_point = self.deproxy.add_endpoint(('localhost',
-                                                    self.deproxy_port))
-        self.handler = deproxy.route('http', 'github.com', self.deproxy)
+        self.end_point = self.deproxy.add_endpoint(self.deproxy_port)
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()
 
     def test_route(self):
-        mc = self.deproxy.make_request('http://localhost:%i/izrik/deproxy' %
-                                       self.deproxy_port)
+        handler = deproxy.route('http', 'httpbin.org', self.deproxy)
+        mc = self.deproxy.make_request('http://localhost:%i/' %
+                                       self.deproxy_port,
+                                       default_handler=handler)
         self.assertEquals(int(mc.received_response.code), 200)
 
 
@@ -105,8 +102,7 @@ class TestCustomHandlers(unittest.TestCase):
     def setUp(self):
         self.deproxy_port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.end_point = self.deproxy.add_endpoint(('localhost',
-                                                    self.deproxy_port))
+        self.end_point = self.deproxy.add_endpoint(self.deproxy_port)
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()
@@ -118,7 +114,7 @@ class TestCustomHandlers(unittest.TestCase):
                                     body='Snape Kills Dumbledore')
         mc = self.deproxy.make_request('http://localhost:%i/' %
                                        self.deproxy_port,
-                                       handler_function=custom_handler)
+                                       default_handler=custom_handler)
         self.assertEquals(int(mc.received_response.code), 606)
 
     def handler_method(self, request):
@@ -129,23 +125,97 @@ class TestCustomHandlers(unittest.TestCase):
     def test_custom_handler_method(self):
         mc = self.deproxy.make_request('http://localhost:%i/' %
                                        self.deproxy_port,
-                                       handler_function=self.handler_method)
+                                       default_handler=self.handler_method)
         self.assertEquals(int(mc.received_response.code), 606)
+
+
+class TestEndpointDefaultHandler(unittest.TestCase):
+    def setUp(self):
+        self.port = get_next_deproxy_port()
+        self.deproxy = deproxy.Deproxy()
+
+    def test_endpoint_default_handler_function(self):
+        def custom_handler(request):
+            return deproxy.Response(code='601', message='Custom', headers={},
+                                    body=None)
+        self.deproxy.add_endpoint(port=self.port,
+                                  default_handler=custom_handler)
+        url = 'http://localhost:{0}/'.format(self.port)
+        mc = self.deproxy.make_request(url=url)
+
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '601')
+        self.assertEqual(mc.received_response.code, '601')
+
+    def custom_handler_method(self, request):
+        return deproxy.Response(code='602', message='Custom', headers={},
+                                body=None)
+
+    def test_endpoint_default_handler_method(self):
+        self.deproxy.add_endpoint(port=self.port,
+                                  default_handler=self.custom_handler_method)
+        url = 'http://localhost:{0}/'.format(self.port)
+        mc = self.deproxy.make_request(url=url)
+
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '602')
+        self.assertEqual(mc.received_response.code, '602')
+
+    def tearDown(self):
+        self.deproxy.shutdown_all_endpoints()
+
+
+class TestDeproxyDefaultHandler(unittest.TestCase):
+    def setUp(self):
+        self.port = get_next_deproxy_port()
+
+    def test_deproxy_default_handler_function(self):
+        def custom_handler(request):
+            return deproxy.Response(code='603', message='Custom', headers={},
+                                    body=None)
+        self.deproxy = deproxy.Deproxy(default_handler=custom_handler)
+        self.deproxy.add_endpoint(port=self.port)
+
+        url = 'http://localhost:{0}/'.format(self.port)
+        mc = self.deproxy.make_request(url=url)
+
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '603')
+        self.assertEqual(mc.received_response.code, '603')
+
+    def custom_handler_method(self, request):
+        return deproxy.Response(code='604', message='Custom', headers={},
+                                body=None)
+
+    def test_deproxy_default_handler_method(self):
+        self.deproxy = deproxy.Deproxy(
+            default_handler=self.custom_handler_method)
+        self.deproxy.add_endpoint(port=self.port)
+
+        url = 'http://localhost:{0}/'.format(self.port)
+        mc = self.deproxy.make_request(url=url)
+
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '604')
+        self.assertEqual(mc.received_response.code, '604')
+
+    def tearDown(self):
+        if hasattr(self, 'deproxy') and self.deproxy is not None:
+            self.deproxy.shutdown_all_endpoints()
 
 
 class TestOrphanedHandlings(unittest.TestCase):
     def setUp(self):
         self.deproxy_port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.end_point = self.deproxy.add_endpoint(('localhost',
-                                                    self.deproxy_port))
+        self.end_point = self.deproxy.add_endpoint(self.deproxy_port)
         self.other_client = deproxy.Deproxy()
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()
 
     def test_orphaned_handling(self):
-        delayed_handler = deproxy.delay(2, deproxy.default_handler)
+        delayed_handler = deproxy.delay(2, deproxy.simple_handler)
         self.long_running_mc = None
 
         class Helper:
@@ -156,7 +226,7 @@ class TestOrphanedHandlings(unittest.TestCase):
         def other_thread():
             mc = self.deproxy.make_request('http://localhost:%i/' %
                                            self.deproxy_port,
-                                           handler_function=delayed_handler)
+                                           default_handler=delayed_handler)
             helper.mc = mc
 
         t = threading.Thread(target=other_thread)
@@ -175,13 +245,13 @@ class TestEndpointShutdown(unittest.TestCase):
         self.deproxy = deproxy.Deproxy()
 
     def test_shutdown(self):
-        e1 = self.deproxy.add_endpoint(('localhost', self.deproxy_port1))
-        e2 = self.deproxy.add_endpoint(('localhost', self.deproxy_port2))
+        e1 = self.deproxy.add_endpoint(self.deproxy_port1)
+        e2 = self.deproxy.add_endpoint(self.deproxy_port2)
 
         e1.shutdown()
 
         try:
-            e3 = self.deproxy.add_endpoint(('localhost', self.deproxy_port1))
+            e3 = self.deproxy.add_endpoint(self.deproxy_port1)
         except socket.error as e:
             self.fail('Address already in use: %s' % e)
 
@@ -193,18 +263,18 @@ class TestShutdownAllEndpoints(unittest.TestCase):
         self.deproxy = deproxy.Deproxy()
 
     def test_shutdown(self):
-        e1 = self.deproxy.add_endpoint(('localhost', self.deproxy_port1))
-        e2 = self.deproxy.add_endpoint(('localhost', self.deproxy_port2))
+        e1 = self.deproxy.add_endpoint(self.deproxy_port1)
+        e2 = self.deproxy.add_endpoint(self.deproxy_port2)
 
         self.deproxy.shutdown_all_endpoints()
 
         try:
-            e3 = self.deproxy.add_endpoint(('localhost', self.deproxy_port1))
+            e3 = self.deproxy.add_endpoint(self.deproxy_port1)
         except socket.error as e:
             self.fail('add_endpoint through an exception: %s' % e)
 
         try:
-            e4 = self.deproxy.add_endpoint(('localhost', self.deproxy_port2))
+            e4 = self.deproxy.add_endpoint(self.deproxy_port2)
         except socket.error as e:
             self.fail('add_endpoint through an exception: %s' % e)
 
@@ -213,7 +283,7 @@ class TestAutomaticRequestHeaders(unittest.TestCase):
     def setUp(self):
         self.port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.endpoint = self.deproxy.add_endpoint(('localhost', self.port))
+        self.endpoint = self.deproxy.add_endpoint(self.port)
         self.url = 'http://localhost:{}/'.format(self.port)
 
     def tearDown(self):
@@ -250,7 +320,7 @@ class TestDefaultResponseHeaders(unittest.TestCase):
     def setUpClass(self):
         self.port = get_next_deproxy_port()
         self.deproxy = deproxy.Deproxy()
-        self.endpoint = self.deproxy.add_endpoint(('localhost', self.port))
+        self.endpoint = self.deproxy.add_endpoint(self.port)
         self.url = 'http://localhost:{}/'.format(self.port)
 
     @classmethod
@@ -275,7 +345,7 @@ class TestDefaultResponseHeaders(unittest.TestCase):
 
     def test_not_specified(self):
         mc = self.deproxy.make_request(url=self.url,
-                                       handler_function=self.handler1)
+                                       default_handler=self.handler1)
         self.assertEqual(len(mc.handlings), 1)
         self.assertIn('server', mc.received_response.headers)
         self.assertIn('date', mc.received_response.headers)
@@ -284,7 +354,7 @@ class TestDefaultResponseHeaders(unittest.TestCase):
 
     def test_explicit_on(self):
         mc = self.deproxy.make_request(url=self.url,
-                                       handler_function=self.handler2)
+                                       default_handler=self.handler2)
         self.assertEqual(len(mc.handlings), 1)
         self.assertIn('server', mc.received_response.headers)
         self.assertIn('date', mc.received_response.headers)
@@ -293,7 +363,7 @@ class TestDefaultResponseHeaders(unittest.TestCase):
 
     def test_explicit_off(self):
         mc = self.deproxy.make_request(url=self.url,
-                                       handler_function=self.handler3)
+                                       default_handler=self.handler3)
         self.assertEqual(len(mc.handlings), 1)
         self.assertNotIn('server', mc.received_response.headers)
         self.assertNotIn('date', mc.received_response.headers)
@@ -359,7 +429,7 @@ class TestBodies(unittest.TestCase):
     def setUp(self):
         self.deproxy = deproxy.Deproxy()
         self.port = get_next_deproxy_port()
-        self.deproxy.add_endpoint(('localhost', self.port))
+        self.deproxy.add_endpoint(self.port)
         self.url = 'http://localhost:{0}/'.format(self.port)
 
     def test_request_body(self):
@@ -383,7 +453,7 @@ class TestBodies(unittest.TestCase):
             return deproxy.Response(code=200, message='OK', headers=None,
                                     body=body)
         mc = self.deproxy.make_request(url=self.url,
-                                       handler_function=custom_handler)
+                                       default_handler=custom_handler)
         self.assertEqual(mc.received_response.body, body)
         self.assertEqual(len(mc.handlings), 1)
         self.assertEqual(mc.handlings[0].response.body, body)
@@ -404,7 +474,7 @@ class TestSendingHeaders(unittest.TestCase):
     def setUp(self):
         self.deproxy = deproxy.Deproxy()
         self.port = get_next_deproxy_port()
-        self.deproxy.add_endpoint(('localhost', self.port))
+        self.deproxy.add_endpoint(self.port)
         self.url = 'http://localhost:{0}/'.format(self.port)
 
     def test_send_duplicate_request_headers(self):
@@ -426,12 +496,98 @@ class TestSendingHeaders(unittest.TestCase):
                                     body=None)
 
         mc = self.deproxy.make_request(url=self.url,
-                                       handler_function=custom_handler)
+                                       default_handler=custom_handler)
 
         self.assertEqual(len(mc.handlings), 1)
         values = [value for value in
                   mc.received_response.headers.find_all('Name')]
         self.assertEqual(values, ['Value1', 'Value2'])
+
+    def tearDown(self):
+        self.deproxy.shutdown_all_endpoints()
+
+
+class TestPerEndpointHandlers(unittest.TestCase):
+    def setUp(self):
+        self.deproxy = deproxy.Deproxy()
+        self.endpoint1 = self.deproxy.add_endpoint(
+            name='test-endpoint-1',
+            port=get_next_deproxy_port())
+        self.endpoint2 = self.deproxy.add_endpoint(
+            name='test-endpoint-2',
+            port=get_next_deproxy_port())
+
+        def custom_handler1(request):
+            return deproxy.Response(code='605', message='Custom', headers={},
+                                    body=None)
+
+        def custom_handler2(request):
+            return deproxy.Response(code='606', message='Spoiler', headers={},
+                                    body=None)
+        self.custom_handler1 = custom_handler1
+        self.custom_handler2 = custom_handler2
+        self.url1 = 'http://localhost:{0}/'.format(self.endpoint1.port)
+        self.url2 = 'http://localhost:{0}/'.format(self.endpoint2.port)
+
+    def test_no_handlers(self):
+        mc = self.deproxy.make_request(url=self.url1)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '200')
+        self.assertEqual(mc.received_response.code, '200')
+
+        mc = self.deproxy.make_request(url=self.url2)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '200')
+        self.assertEqual(mc.received_response.code, '200')
+
+    def test_empty_handlers(self):
+        mc = self.deproxy.make_request(url=self.url1, handlers={})
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '200')
+        self.assertEqual(mc.received_response.code, '200')
+
+        mc = self.deproxy.make_request(url=self.url2, handlers={})
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '200')
+        self.assertEqual(mc.received_response.code, '200')
+
+    def test_both_handlers(self):
+        handlers = {self.endpoint1: self.custom_handler1,
+                    self.endpoint2: self.custom_handler2}
+        mc = self.deproxy.make_request(url=self.url1, handlers=handlers)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '605')
+        self.assertEqual(mc.received_response.code, '605')
+
+        mc = self.deproxy.make_request(url=self.url2, handlers=handlers)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '606')
+        self.assertEqual(mc.received_response.code, '606')
+
+    def test_one_handler(self):
+        handlers = {self.endpoint1: self.custom_handler1}
+        mc = self.deproxy.make_request(url=self.url1, handlers=handlers)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '605')
+        self.assertEqual(mc.received_response.code, '605')
+
+        mc = self.deproxy.make_request(url=self.url2, handlers=handlers)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '200')
+        self.assertEqual(mc.received_response.code, '200')
+
+    def test_handlers_by_name(self):
+        handlers = {'test-endpoint-1': self.custom_handler1,
+                    'test-endpoint-2': self.custom_handler2}
+        mc = self.deproxy.make_request(url=self.url1, handlers=handlers)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '605')
+        self.assertEqual(mc.received_response.code, '605')
+
+        mc = self.deproxy.make_request(url=self.url2, handlers=handlers)
+        self.assertEqual(len(mc.handlings), 1)
+        self.assertEqual(mc.handlings[0].response.code, '606')
+        self.assertEqual(mc.received_response.code, '606')
 
     def tearDown(self):
         self.deproxy.shutdown_all_endpoints()